+++ device-bootloaders/CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-17565] Fixed sending block of pmic fw update data and increased number of bytes per transaction
 - [SW-17727] Fixed issue with Throttle down  status not being updated correcltly
 - [SW-17906] Fixed power throtting to properly handle errors retuned form function calls
+  - Fixed checking status of operating point change in power throttle 
 ### Security
 
 ## [0.12.0] - 2023-06-02